# Conflicts:
#	src/main/java/iriro/publicData/repository/CrimeRoadRepository.java
#	src/main/java/iriro/publicData/repository/FacilitySafeRepository.java

+++ src/main/java/iriro/saferoute/dto/DetourWayPointDto.java
@@ -5,10 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor@NoArgsConstructor@Data@Builder
 public class DetourWayPointDto {
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
 }
