+++ src/main/java/iriro/saferoute/dto/RoutePointDto.java
@@ -6,14 +6,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 @Builder
 public class RoutePointDto {
-    private BigDecimal latitude; // 위도
-    private BigDecimal longitude; // 경도
+    private Double latitude; // 위도
+    private Double longitude; // 경도
     private Integer sequence; // 경로 순서 -> 몇 번째 지점이 위험구역인지 판단
 }
