+++ src/main/java/iriro/saferoute/dto/RouteResponseDto.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @NoArgsConstructor
@@ -13,10 +12,10 @@ import java.util.List;
 @Data
 @Builder
 public class RouteResponseDto {
-    private BigDecimal start_latitude;
-    private BigDecimal start_longitude;
-    private BigDecimal end_latitude;
-    private BigDecimal end_longitude;
+    private Double start_latitude;
+    private Double start_longitude;
+    private Double end_latitude;
+    private Double end_longitude;
 
     private Integer totalTime;
     private Integer totalDistance;
