+++ src/main/java/iriro/saferoute/dto/SafetyFacPointDto.java
@@ -5,12 +5,10 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor @NoArgsConstructor @Data @Builder
 public class SafetyFacPointDto {
     private String facType; // 안전시설물 타입
     private int safeCount; // 안전 시설물 개수
-    private BigDecimal latitude; // 위도
-    private BigDecimal longitude; // 경도
+    private Double latitude; // 위도
+    private Double longitude; // 경도
 }
